add validation to dto object members

+++ src/main/java/com/example/demo/services/impl/AgendaService.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 /**
  * @author Okala III
@@ -24,9 +25,20 @@ public class AgendaService implements CRUDService<Agenda, AgendaDto> {
 
     @Override
     public Agenda create(AgendaDto dto) {
-        //todo validate objects
+
+        if (dto.getDate().toString().isBlank())
+            throw new IllegalArgumentException("Invalid Agenda Date");
+
+        if (dto.getDetails().isBlank())
+            throw new IllegalArgumentException("Details field can not be empty");
+
+        if (dto.getName().isBlank())
+            throw new IllegalArgumentException("Name field can not be empty");
+
+        if (dto.getTitle().isBlank())
+            throw new IllegalArgumentException("title field can not be empty");
         agenda = new Agenda();
-        BeanUtils.copyProperties(dto,agenda);
+        BeanUtils.copyProperties(dto, agenda);
 
         Agenda savedAgenda = agendaRepository.save(agenda);
         return savedAgenda;
@@ -34,30 +46,49 @@ public class AgendaService implements CRUDService<Agenda, AgendaDto> {
     }
 
     @Override
-    public List<Agenda> create(List<AgendaDto> d) {
+    public List<Agenda> create(List<AgendaDto> dto) {
         return null;
     }
 
     @Override
-    public Agenda update(AgendaDto d) {
-        //todo validate objects
-        agenda = agendaRepository.findById(d.getId()).orElseThrow(()-> new IllegalArgumentException("Invalid Agenda Id"));
-        BeanUtils.copyProperties(d,agenda);
+    public Agenda update(AgendaDto dto) {
+        if (dto.getId() == null || dto.getId() == 0)
+            throw new IllegalArgumentException("Invalid Agenda Id");
+        
+        if (!agendaRepository.existsById(dto.getId())) 
+            throw new NoSuchElementException("Invalid Agenda ID");
+        
+        if (dto.getDate().toString().isBlank())
+            throw new IllegalArgumentException("Invalid Agenda Date");
+
+        if (dto.getDetails().isBlank())
+            throw new IllegalArgumentException("Details field can not be empty");
+
+        if (dto.getName().isBlank())
+            throw new IllegalArgumentException("Name field can not be empty");
+
+        if (dto.getTitle().isBlank())
+            throw new IllegalArgumentException("title field can not be empty");
+
+        agenda = agendaRepository.findById(dto.getId()).orElseThrow(() -> new IllegalArgumentException("Invalid Agenda Id"));
+        BeanUtils.copyProperties(dto, agenda);
 
         Agenda updatedAgenda = agendaRepository.save(agenda);
         return updatedAgenda;
     }
 
     @Override
-    public List<Agenda> update(List<AgendaDto> d) {
+    public List<Agenda> update(List<AgendaDto> dto) {
         return null;
     }
 
     @Override
     public Boolean delete(Long id) {
-            // todo check if Id exist
-    agendaRepository.deleteById(id);
-    return true;
+        if (!agendaRepository.existsById(id)) {
+            throw new NoSuchElementException("Invalid Agenda Id");
+        }
+        agendaRepository.deleteById(id);
+        return true;
     }
 
     @Override
@@ -67,7 +98,12 @@ public class AgendaService implements CRUDService<Agenda, AgendaDto> {
 
     @Override
     public Agenda findById(long id) {
-        return agendaRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("invalid Agenda Id"));
+        if (id <= 0 )
+            throw new IllegalArgumentException("Invalid Agenda Id");
+
+        if (!agendaRepository.existsById(id))
+            throw new NoSuchElementException("Invalid Agenda ID");
+        return agendaRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("invalid Agenda Id"));
     }
 
     public List<Agenda> findAllCompleted() {
